requirements.txt
beautifulsoup4
selenium
load_dotenv # 加载环境变量